requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pydantic==1.10.2
google-generativeai
requests
fastapi[all]
